rtc: mpfs: fix datetime register read order
The read order of the datetime registers is incorrect. The hardware will
latch the registers so that a read from the upper register *after* the
lower register will always contain the upper value aligned to the last
lower read. Currently the upper register is read first, which may not be
aligned with the lower register.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>