pwm: microchip-core: account for edge "quirk" while inverted
When polarity is inverted, turning off the output is not a sufficient
workaround for being unable to set pos & neg edges equal to each other
as disabling the output sets it low.
Work around this by setting the high time to its maximum possible value
while inverted. The same approach can be used for "normal" polarity, so
for the sake of simplicity convert it to use the same method.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>