pwm: microchip-core: fix percieved idempotency and rounding issues
When shadow registers are enabled in the IP core, the values just
written to the various waveform control registers (including the enable)
do not appear on the bus until the waveform has been updated at the
start of the next period. Add a delay before returning from apply until
the values will be visible so that the PWM core does not perceive
idempotency or rounding problems.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>